requirements.txt
# Kimi Writer MCP Server - Unified Dependencies
# Install with: pip install -r requirements.txt

# Core Kimi Writer dependencies
openai>=1.0.0
httpx>=0.24.0
python-dotenv>=1.0.0

# Model Context Protocol
mcp>=1.0.0

# PDF Generation
weasyprint>=60.0
markdown>=3.5.1
markdown2>=2.4.8

# EPUB Generation  
ebooklib>=0.18
beautifulsoup4>=4.12.0
jinja2>=3.1.0

# Additional utilities for document processing
python-frontmatter>=1.0.0  # For handling markdown metadata
pypandoc>=1.13             # Pandoc wrapper (optional alternative)
html2text>=2020.1.16       # HTML to text conversion

# For development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# For better logging and debugging
rich>=13.0.0

# For data validation (alternative to pydantic that comes with MCP)
# pydantic>=2.0.0